backend/requirements.txt
Django>=5.2
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
django-filter>=23.0
django-solo>=2.0.0
django-redis>=5.0.0
python-decouple>=3.8
python-dotenv>=1.0.0
django-imagekit>=5.0.0
gunicorn>=21.2.0

